requirements.txt
Flask==2.3.3
fpdf2==2.7.4
python-dotenv==1.0.0
gunicorn==21.2.0